backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
faiss-cpu
pdfplumber
python-dotenv
huggingface-hub
sentence-transformers
